destroy

+++ src/workspace/thread.js
@@ -128,13 +128,14 @@ Entry.Thread = function(thread, code, parent) {
     };
 
     p.destroy = function(animate) {
-        this._code.destroyThread(this, false);
         if (this.view) this.view.destroy(animate);
 
         var blocks = this._data;
 
         for (var i=blocks.length-1; i>=0; i--)
             blocks[i].destroy(animate);
+
+        !blocks.length && this._code.destroyThread(this, false);
     };
 
     p.getBlock = function(index) {
